move infrastructure elements connected with domain to domain package
