+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/Customer.java
@@ -2,13 +2,14 @@ package com.asc.loanservice.domain.loan.application;
 
 import lombok.Builder;
 import lombok.Getter;
-
+import javax.persistence.Embeddable;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
 @Getter
 @Builder
+@Embeddable
 public class Customer {
 
     private final String customerName;
