+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/LoanMapper.java
@@ -1,10 +1,14 @@
-package com.asc.loanservice.domain.loan.application.port;
+package com.asc.loanservice.domain.loan.application;
 
 import com.asc.loanservice.domain.loan.application.contract.LoanApplicationCreateInput;
 import com.asc.loanservice.domain.loan.application.contract.LoanApplicationRequest;
+import org.mapstruct.Mapper;
+import org.springframework.stereotype.Component;
 
 
-public interface LoanMapperPort {
+@Mapper
+@Component
+interface LoanMapper {
 
     LoanApplicationCreateInput createInput(LoanApplicationRequest loanApplicationRequest);
 }
