--- src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/port/LoanApplicationDataProviderPort.java
@@ -1,12 +1,0 @@
-package com.asc.loanservice.domain.loan.application.port;
-
-import com.asc.loanservice.infrastructure.repository.model.LoanApplication;
-
-import java.util.Optional;
-
-
-public interface LoanApplicationDataProviderPort {
-
-    LoanApplication save(LoanApplication loanApplication);
-    Optional<LoanApplication> findByLoanRequestNumber(String loanRequestNumber);
-}
